refactor: #35 태그 enums 패키지 이동
