+++ src/main/java/com/dnd/MusicLog/log/dto/SaveLogRequestDto.java
@@ -1,10 +1,10 @@
 package com.dnd.MusicLog.log.dto;
 
 import com.dnd.MusicLog.music.enums.MusicType;
-import com.dnd.MusicLog.taginfo.enums.Feeling;
-import com.dnd.MusicLog.taginfo.enums.Season;
-import com.dnd.MusicLog.taginfo.enums.Time;
-import com.dnd.MusicLog.taginfo.enums.Weather;
+import com.dnd.MusicLog.log.enums.Feeling;
+import com.dnd.MusicLog.log.enums.Season;
+import com.dnd.MusicLog.log.enums.Time;
+import com.dnd.MusicLog.log.enums.Weather;
 import jakarta.annotation.Nullable;
 
 import java.time.LocalDate;
