+++ src/main/java/com/dnd/MusicLog/taginfo/dto/Tag.java
@@ -1,9 +1,9 @@
 package com.dnd.MusicLog.taginfo.dto;
 
-import com.dnd.MusicLog.taginfo.enums.Feeling;
-import com.dnd.MusicLog.taginfo.enums.Season;
-import com.dnd.MusicLog.taginfo.enums.Time;
-import com.dnd.MusicLog.taginfo.enums.Weather;
+import com.dnd.MusicLog.log.enums.Feeling;
+import com.dnd.MusicLog.log.enums.Season;
+import com.dnd.MusicLog.log.enums.Time;
+import com.dnd.MusicLog.log.enums.Weather;
 
 public record Tag(Feeling feeling, Time time, Weather weather, Season season) {
 }
